requirements.txt
feedparser
requests
markdown
pyyaml